gps datatype fix

+++ End Device/gps-geiger/container.h
@@ -5,8 +5,8 @@
 #include "Time.h"
 
 struct Container {
-    double latitude = 0;
-    double longitude = 0;
+    int32_t latitude = 0;
+    int32_t longitude = 0;
     time_t time = 0;
-    float geiger = 0;
+    int32_t geiger = 0;
 };
